feat(elections): always use curr eff stake instead of min(curr, prev)
- split runner & strategy tests
