+++ src/node-control/elections/src/runner_tests.rs
@@ -7,7 +7,6 @@
  * This software is provided "AS IS", WITHOUT WARRANTY OF ANY KIND.
  */
 use super::*;
-use crate::adaptive_strategy::calc_adaptive_stake;
 use common::{
     app_config::{ElectionsConfig, NodeBinding, StakePolicy},
     snapshot::SnapshotStore,
@@ -1931,339 +1930,8 @@ fn test_compute_status_idle_when_enabled_no_recover_no_participant() {
     assert_eq!(status, BindingStatus::Idle);
 }
 
-// =====================================================
-// AdaptiveSplit50: calc_adaptive_stake unit tests
-// =====================================================
-
 const NANO: u64 = 1_000_000_000;
 const FACTOR_3X: u32 = 3 * 65536;
-
-/// Build an ElectionsInfo with `n` participants each staking `stake_per` nanotons.
-fn elections_info_with_participants(n: usize, stake_per: u64) -> ElectionsInfo {
-    let participants = (0..n)
-        .map(|i| {
-            let mut pubkey = [0u8; 32];
-            pubkey[0] = i as u8;
-            pubkey[1] = (i >> 8) as u8;
-            Participant {
-                pub_key: pubkey.to_vec(),
-                adnl_addr: [0xEE; 32].to_vec(),
-                wallet_addr: pubkey.to_vec(),
-                stake: stake_per,
-                max_factor: FACTOR_3X,
-                election_id: ELECTION_ID,
-                stake_message_boc: None,
-            }
-        })
-        .collect();
-    ElectionsInfo {
-        election_id: ELECTION_ID,
-        elect_close: ELECTION_ID + 600,
-        min_stake: MIN_STAKE,
-        total_stake: n as u64 * stake_per,
-        failed: false,
-        finished: false,
-        participants,
-    }
-}
-
-// ---- Step 3.4: half >= min_eff → stake half ----
-
-#[test]
-fn test_adaptive_stake_half_when_above_min_eff() {
-    // 50 participants with 300k TON each, max_validators=400 (set NOT full).
-    // effective_min is ~100k TON (300k / factor 3).
-    // total_balance = 1_300_000 TON, half = 650_000 TON >> effective_min.
-    // Expected: stake half = 650_000 TON.
-    let total_balance = 1_300_000 * NANO;
-    let free_balance = total_balance; // no frozen, no current
-    let current_stake = 0;
-
-    let elections_info = elections_info_with_participants(50, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        None,
-    )
-    .unwrap();
-
-    let half = total_balance / 2;
-    assert_eq!(result, half, "should stake half");
-}
-
-// ---- Step 3.5: half < min_eff → stake all ----
-
-#[test]
-fn test_adaptive_stake_all_when_half_below_min_eff() {
-    // 400 participants with ~700k TON each (set FULL, max_validators=400).
-    // effective_min ~700k TON. Our total = 1_300_000 TON, half = 650_000 < 700_000.
-    // Expected: stake all free_balance.
-    let total_balance = 1_300_000 * NANO;
-    let free_balance = total_balance;
-    let current_stake = 0;
-
-    let stakes: Vec<Participant> = (0..400)
-        .map(|i| {
-            let mut pubkey = [0u8; 32];
-            pubkey[0] = i as u8;
-            pubkey[1] = (i >> 8) as u8;
-            Participant {
-                pub_key: pubkey.to_vec(),
-                adnl_addr: [0xEE; 32].to_vec(),
-                wallet_addr: pubkey.to_vec(),
-                stake: (700_000 + i as u64 * 100) * NANO,
-                max_factor: FACTOR_3X,
-                election_id: ELECTION_ID,
-                stake_message_boc: None,
-            }
-        })
-        .collect();
-    let elections_info = ElectionsInfo {
-        election_id: ELECTION_ID,
-        elect_close: ELECTION_ID + 600,
-        min_stake: MIN_STAKE,
-        total_stake: stakes.iter().map(|p| p.stake).sum(),
-        failed: false,
-        finished: false,
-        participants: stakes,
-    };
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        None,
-    )
-    .unwrap();
-
-    assert_eq!(result, free_balance, "should stake all free_balance when half < min_eff");
-}
-
-// ---- Step 4.1: current_stake >= min_eff → no top-up ----
-
-#[test]
-fn test_adaptive_no_topup_when_stake_sufficient() {
-    // 50 participants with 300k TON. effective_min ~100k.
-    // current_stake = 650_000 TON >> effective_min.
-    // Expected: return 0 (no top-up).
-    let total_balance = 1_300_000 * NANO;
-    let free_balance = 0; // all staked or frozen
-    let current_stake = 650_000 * NANO;
-
-    let elections_info = elections_info_with_participants(50, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        None,
-    )
-    .unwrap();
-
-    assert_eq!(result, 0, "should return 0 when current_stake >= min_eff");
-}
-
-// ---- Step 3.5: insufficient funds guard ----
-
-#[test]
-fn test_adaptive_skip_when_insufficient_funds() {
-    // 50 participants with 300k TON. effective_min ~100k.
-    // total_balance is high (due to frozen), but free_balance < required.
-    // Expected: return 0 (skip).
-    let frozen = 900_000 * NANO;
-    let free_balance = 50_000 * NANO; // less than effective_min (~100k)
-    let current_stake = 0;
-    let total_balance = frozen + free_balance + current_stake;
-
-    let elections_info = elections_info_with_participants(50, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        None,
-    )
-    .unwrap();
-
-    assert_eq!(result, 0, "should skip when free_balance < min_eff_stake");
-}
-
-// ---- Cap to free_balance when half > free_balance ----
-
-#[test]
-fn test_adaptive_skip_when_half_exceeds_free_balance() {
-    // Use few participants (< min_validators) so emulation returns None.
-    // prev_min_eff = 50k controls the effective_min.
-    // total = 1_300_000, half = 650_000 > 50k → half branch.
-    // free_balance = 200_000 < half(650k) → skip (not enough to stake half).
-    // free_balance (200k) > prev_min_eff (50k) → passes insufficient funds guard.
-    let frozen = 1_100_000 * NANO;
-    let free_balance = 200_000 * NANO;
-    let current_stake = 0;
-    let total_balance = frozen + free_balance + current_stake;
-    let prev_min_eff = Some(50_000 * NANO);
-
-    let elections_info = elections_info_with_participants(5, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        prev_min_eff,
-    )
-    .unwrap();
-
-    assert_eq!(result, 0, "should skip when free_balance < half (operator should top up pool)");
-}
-
-// ---- min(curr, prev) selection ----
-
-#[test]
-fn test_adaptive_uses_min_of_curr_and_prev() {
-    // 50 participants with 300k TON. curr_min_eff ~100k.
-    // prev_min_eff = 80k < curr → should use prev (80k).
-    // total = 200k, half = 100k >= prev_min_eff (80k) → stake half = 100k.
-    let total_balance = 200_000 * NANO;
-    let free_balance = total_balance;
-    let current_stake = 0;
-    let prev_min_eff = Some(80_000 * NANO);
-
-    let elections_info = elections_info_with_participants(50, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        prev_min_eff,
-    )
-    .unwrap();
-
-    let half = total_balance / 2;
-    assert_eq!(result, half, "should use min(curr, prev) and stake half");
-}
-
-// ---- prev only (curr = None, fewer than min_validators) ----
-
-#[test]
-fn test_adaptive_fallback_to_prev_when_not_enough_participants() {
-    // Only 5 participants (< min_validators=13) → emulation returns None.
-    // prev_min_eff = 50k.
-    // total = 200k, half = 100k >= 50k → stake half.
-    let total_balance = 200_000 * NANO;
-    let free_balance = total_balance;
-    let current_stake = 0;
-    let prev_min_eff = Some(50_000 * NANO);
-
-    let elections_info = elections_info_with_participants(5, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        prev_min_eff,
-    )
-    .unwrap();
-
-    let half = total_balance / 2;
-    assert_eq!(result, half, "should fallback to prev_min_eff when not enough participants");
-}
-
-// ---- Both None → error ----
-
-#[test]
-fn test_adaptive_error_when_no_min_eff_available() {
-    // Fewer than min_validators (5 < 13) AND no prev_min_eff → error.
-    let total_balance = 200_000 * NANO;
-    let free_balance = total_balance;
-    let current_stake = 0;
-
-    let elections_info = elections_info_with_participants(5, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        None,
-    );
-
-    assert!(result.is_err(), "should fail when both curr and prev min_eff are unavailable");
-}
-
-// ---- Top-up: half branch, partial top-up ----
-
-#[test]
-fn test_adaptive_topup_to_half() {
-    // Use few participants so emulation returns None; prev_min_eff controls effective.
-    // prev_min_eff = 600k. current_stake = 500k < 600k → need top-up.
-    // total = 1_300_000, half = 650_000 > 600k → half branch.
-    // stake = half - current = 650k - 500k = 150k.
-    let total_balance = 1_300_000 * NANO;
-    let free_balance = 200_000 * NANO;
-    let current_stake = 500_000 * NANO;
-    let prev_min_eff = Some(600_000 * NANO);
-
-    // current_stake > 0 → emulation uses our_stake = 0 (already in list).
-    // With < min_validators participants, emulation returns None → uses prev_min_eff.
-    let elections_info = elections_info_with_participants(5, 300_000 * NANO);
-
-    let result = calc_adaptive_stake(
-        "node-1",
-        total_balance,
-        free_balance,
-        current_stake,
-        FACTOR_3X,
-        &elections_info,
-        &default_cfg16(),
-        &default_cfg17(),
-        prev_min_eff,
-    )
-    .unwrap();
-
-    let expected = total_balance / 2 - current_stake;
-    assert_eq!(result, expected, "should top up to half");
-}
 
 // =====================================================
 // AdaptiveSplit50: wait/sleep integration tests
